+++ CHANGELOG.md
@@ -6,13 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.0.19] - TBD
 
-Uses libvips v8.19.0, compiled with Emscripten v6.0.6.
+Uses libvips v8.19.0, compiled with Emscripten v6.0.8.
 
 ### Added
 
 - Support for growable WebAssembly memory up to 2 GiB.
   [#106](https://github.com/kleisauke/wasm-vips/issues/106)
 - Support for `workaroundCors` setting in ES6 module.
+
+### Fixed
+
+- Validate typed array format in `Image.newFromMemory()`.
+  [#126](https://github.com/kleisauke/wasm-vips/issues/126)
 
 ### Changed